+++ asilib/array/lamp_phantom.py
@@ -9,14 +9,13 @@ import h5py
 import asilib
 import asilib.utils as utils
 import asilib.io.download as download
-import asilib.io.manager as manager
 
 
 image_base_url = 'https://ergsc.isee.nagoya-u.ac.jp/psa-pwing/pub/raw/lamp/geoff/'
 skymap_base_url = 'https://ergsc.isee.nagoya-u.ac.jp/psa-pwing/pub/raw/lamp/geoff/'
 local_base_dir = asilib.config['ASI_DATA_DIR'] / 'lamp_phantom'
 
-def lamp(location_code, time=None, time_range=None, overwrite=False, missing_ok=True, alt=100):
+def lamp(location_code, time=None, time_range=None, redownload=False, missing_ok=True, alt=100):
     """
     Create an Imager instance using the LAMP's ground-based Phantom ASI.
 
@@ -30,7 +29,7 @@ def lamp(location_code, time=None, time_range=None, overwrite=False, missing_ok=
     time_range: list
         A length 2 list of string-formatted times or datetimes to bracket
         the ASI data time interval.
-    overwrite: bool
+    redownload: bool
         If True, will download the data from the internet, regardless of
         wether or not the data exists locally (useful if the data becomes
         corrupted).
@@ -52,7 +51,7 @@ def lamp(location_code, time=None, time_range=None, overwrite=False, missing_ok=
     else:
         raise NotImplementedError
 
-    _skymap = load_skymap(location_code, alt, overwrite)
+    _skymap = load_skymap(location_code, alt, redownload)
     skymap = {
         'lat':_skymap['Lat'],
         'lon':_skymap['Lon'],
@@ -62,7 +61,7 @@ def lamp(location_code, time=None, time_range=None, overwrite=False, missing_ok=
         'path':_skymap['path']
         }
 
-    file_paths = _get_files(location_code, time, time_range, overwrite, missing_ok)
+    file_paths = _get_files(location_code, time, time_range, redownload, missing_ok)
 
     if time_range is not None:  # Prepare the loader for multiple files
         start_times = len(file_paths)*[None]
@@ -94,7 +93,7 @@ def lamp(location_code, time=None, time_range=None, overwrite=False, missing_ok=
     return asilib.Imager(data, meta, skymap)
 
 
-def _get_files(location_code, time, time_range, overwrite, missing_ok):
+def _get_files(location_code, time, time_range, redownload, missing_ok):
     """
     Get the LAMP Phantom image file paths.
     """
@@ -104,15 +103,18 @@ def _get_files(location_code, time, time_range, overwrite, missing_ok):
         raise ValueError('both time and time_range can not be simultaneously specified.')
 
     local_dir = local_base_dir / 'images' / location_code.lower()
-    _manager = manager.File_Manager(local_dir, base_url=image_base_url)
-
-    # Find one image file.
+    
     if time is not None:
         time = utils.validate_time(time)
-        # # file_paths will only contain one path.
         filename = f'narrow_{time.strftime("%H%M")}.mat'
-        file_paths = [_manager.find_file(filename, overwrite=overwrite)]
-        
+        file_paths = list(pathlib.Path(local_dir).rglob(filename))
+
+        if (len(file_paths) == 0) or redownload:
+            d = download.Downloader(image_base_url + f'{filename}')
+            file_paths = d.download(local_dir, redownload=redownload, stream=True)
+        else:
+            return file_paths
+
     # Find multiple image files.
     if time_range is not None:
         time_range = utils.validate_time_range(time_range)
@@ -121,18 +123,20 @@ def _get_files(location_code, time, time_range, overwrite, missing_ok):
 
         for file_time in file_times:
             filename = f'narrow_{file_time.strftime("%H%M")}.mat'
-            try:
-                file_paths.append(
-                    _manager.find_file(filename, overwrite=overwrite)
-                )
-            except (FileNotFoundError, AssertionError) as err:
-                if (missing_ok and 
-                    (
-                        ('does not contain any hyper references containing' in str(err)) or
-                        ('Only one href is allowed' in str(err))
-                    )):
-                    continue
-                else:
+            _local_file_path = list(pathlib.Path(local_dir).rglob(filename))
+
+            if (len(_local_file_path) == 0) or redownload:
+                d = download.Downloader(image_base_url + f'{filename}')
+                try:
+                    file_paths.append(d.download(local_dir, redownload=redownload, stream=True))
+                except (FileNotFoundError, AssertionError, ConnectionError) as err:
+                    if (missing_ok and 
+                        (
+                            ('does not contain any hyper references containing' in str(err)) or
+                            ('Only one href is allowed' in str(err)) or
+                            ('error response' in str(err))
+                        )):
+                        continue
                     raise
     return file_paths
 
@@ -161,7 +165,7 @@ def lamp_reader(path):
             # images = f[image_key][pos:pos+chunk_size]
             yield times, images
 
-def find_skymap(location_code, alt, overwrite=True):
+def find_skymap(location_code, alt, redownload=True):
     """
     Find the path to the skymap file.
     """
@@ -170,7 +174,7 @@ def find_skymap(location_code, alt, overwrite=True):
     # Check if the AzEl skymap is already downloaded.
     local_azel_paths = list(pathlib.Path(local_dir).rglob('Average_fps_im_-17700_to_-17600_AzEl.txt'))
 
-    if (len(local_azel_paths) == 0) or overwrite:
+    if (len(local_azel_paths) == 0) or redownload:
         # Download the skymaps.
         parent_folders = download.Downloader(skymap_base_url)
         parent_folders.find_url(filename='Average_fps_im_-17700_to_-17600_AzEl.txt')
@@ -186,11 +190,11 @@ def find_skymap(location_code, alt, overwrite=True):
         f'Average_fps_im_-17700_to_-17600_LatLong{alt:03}km.txt'
         ))
 
-    if (len(local_latlon_paths) == 0) or overwrite:
+    if (len(local_latlon_paths) == 0) or redownload:
         # Download the skymaps.
         parent_folders = download.Downloader(skymap_base_url)
         parent_folders.find_url(filename=f'Average_fps_im_-17700_to_-17600_LatLong{alt:03}km.txt')
-        local_latlon_paths = parent_folders.download(local_dir) 
+        local_latlon_paths = parent_folders.download(local_dir, redownload=redownload) 
 
     if len(local_latlon_paths) != 1:
         raise FileNotFoundError(
@@ -199,11 +203,11 @@ def find_skymap(location_code, alt, overwrite=True):
     return local_azel_paths[0], local_latlon_paths[0]
 
 
-def load_skymap(location_code, alt, overwrite):
+def load_skymap(location_code, alt, redownload):
     """
     Load the skymap file and apply the transformations.
     """
-    azel_path, latlon_path = find_skymap(location_code, alt, overwrite)
+    azel_path, latlon_path = find_skymap(location_code, alt, redownload)
     skymap = _load_skymap(azel_path, latlon_path)
     # skymap = _tranform_longitude_to_180(skymap)
     # skymap = _transform_azimuth_to_180(skymap)