+++ asilib/io/download.py
@@ -1,5 +1,7 @@
-from typing import List, Union
+from __future__ import annotations  # to support the -> List[Downloader] return type
+from typing import List
 import pathlib
+import urllib
 import re
 
 from bs4 import BeautifulSoup
@@ -23,7 +25,7 @@ def download_image(
     location_code: str,
     time: utils._time_type = None,
     time_range: utils._time_range_type = None,
-    force_download: bool = False,
+    redownload: bool = False,
     ignore_missing_data: bool = True,
 ) -> List[pathlib.Path]:
     """
@@ -41,7 +43,7 @@ def download_image(
         ISO 8601 standard.
     time_range: list of datetime.datetimes or stings
         Defined the duration of data to download. Must be of length 2.
-    force_download: bool
+    redownload: bool
         If True, download the file even if it already exists. Useful if a prior
         data download was incomplete.
     ignore_missing_data: bool
@@ -72,7 +74,7 @@ def download_image(
             location_code,
             time=time,
             time_range=time_range,
-            force_download=force_download,
+            redownload=redownload,
             ignore_missing_data=ignore_missing_data,
         )
     elif asi_array_code.lower() == 'rego':
@@ -80,14 +82,14 @@ def download_image(
             location_code,
             time=time,
             time_range=time_range,
-            force_download=force_download,
+            redownload=redownload,
             ignore_missing_data=ignore_missing_data,
         )
     return paths
 
 
 def download_skymap(
-    asi_array_code: str, location_code: str, force_download: bool = False
+    asi_array_code: str, location_code: str, redownload: bool = False
 ) -> List[pathlib.Path]:
     """
     Download all of the THEMIS or REGO skymap IDL .sav files.
@@ -98,7 +100,7 @@ def download_skymap(
         The imager array name, i.e. ``THEMIS`` or ``REGO``.
     location_code: str
         The ASI station code, i.e. ``ATHA``
-    force_download: bool
+    redownload: bool
         If True, download the file even if it already exists. Useful if a prior
         data download was incomplete.
 
@@ -117,9 +119,9 @@ def download_skymap(
     | asilib.download_skymap(asi_array_code, location_code)
     """
     if asi_array_code.lower() == 'themis':
-        paths = download_themis_skymap(location_code, force_download=force_download)
+        paths = download_themis_skymap(location_code, redownload=redownload)
     elif asi_array_code.lower() == 'rego':
-        paths = download_rego_skymap(location_code, force_download=force_download)
+        paths = download_rego_skymap(location_code, redownload=redownload)
     return paths
 
 
@@ -127,7 +129,7 @@ def download_themis_img(
     location_code: str,
     time: utils._time_type = None,
     time_range: utils._time_range_type = None,
-    force_download: bool = False,
+    redownload: bool = False,
     ignore_missing_data: bool = True,
 ) -> List[pathlib.Path]:
     """
@@ -143,7 +145,7 @@ def download_themis_img(
         ISO 8601 standard.
     time_range: list of datetime.datetimes or stings
         Defined the duration of data to download. Must be of length 2.
-    force_download: bool
+    redownload: bool
         If True, download the file even if it already exists. Useful if a prior
         data download was incomplete.
     ignore_missing_data: bool
@@ -176,7 +178,7 @@ def download_themis_img(
     elif time is not None:
         time = utils._validate_time(time)
         download_path = _download_one_img_file(
-            'themis', location_code, base_url, time, force_download
+            'themis', location_code, base_url, time, redownload
         )
         download_paths = [
             download_path
@@ -190,7 +192,7 @@ def download_themis_img(
         for hour in download_hours:
             try:
                 download_path = _download_one_img_file(
-                    'themis', location_code, base_url, hour, force_download
+                    'themis', location_code, base_url, hour, redownload
                 )
                 download_paths.append(download_path)
             except NotADirectoryError:
@@ -202,7 +204,7 @@ def download_themis_img(
     return download_paths
 
 
-def download_themis_skymap(location_code: str, force_download: bool = False) -> List[pathlib.Path]:
+def download_themis_skymap(location_code: str, redownload: bool = False) -> List[pathlib.Path]:
     """
     Download all of the THEMIS skymap IDL .sav files.
 
@@ -210,7 +212,7 @@ def download_themis_skymap(location_code: str, force_download: bool = False) -> 
     ----------
     location_code: str
         The ASI station code, i.e. ``ATHA``
-    force_download: bool
+    redownload: bool
         If True, download the file even if it already exists. Useful if a prior
         data download was incomplete.
 
@@ -246,10 +248,10 @@ def download_themis_skymap(location_code: str, force_download: bool = False) -> 
         skymap_name = utils._search_hrefs(skymap_folder_absolute, search_pattern='.sav')[0]
         skymap_save_name = skymap_name.replace('-%2B', '')  # Replace the unicode '+'.
 
-        # Download if force_download=True or the file does not exist.
+        # Download if redownload=True or the file does not exist.
         download_path = pathlib.Path(save_dir, skymap_save_name)
         download_paths.append(download_path)
-        if force_download or (not download_path.is_file()):
+        if redownload or (not download_path.is_file()):
             utils._stream_large_file(skymap_folder_absolute + skymap_name, download_path)
     return download_paths
 
@@ -258,7 +260,7 @@ def download_rego_img(
     location_code: str,
     time: utils._time_type = None,
     time_range: utils._time_range_type = None,
-    force_download: bool = False,
+    redownload: bool = False,
     ignore_missing_data: bool = True,
 ) -> List[pathlib.Path]:
     """
@@ -274,7 +276,7 @@ def download_rego_img(
         ISO 8601 standard.
     time_range: list of datetime.datetimes or stings
         Defined the duration of data to download. Must be of length 2.
-    force_download: bool
+    redownload: bool
         If True, download the file even if it already exists. Useful if a prior
         data download was incomplete.
     ignore_missing_data: bool
@@ -308,7 +310,7 @@ def download_rego_img(
     elif time is not None:
         time = utils._validate_time(time)
         download_path = _download_one_img_file(
-            'rego', location_code, base_url, time, force_download
+            'rego', location_code, base_url, time, redownload
         )
         download_paths = [
             download_path
@@ -322,7 +324,7 @@ def download_rego_img(
         for hour in download_hours:
             try:
                 download_path = _download_one_img_file(
-                    'rego', location_code, base_url, hour, force_download
+                    'rego', location_code, base_url, hour, redownload
                 )
                 download_paths.append(download_path)
             except NotADirectoryError:
@@ -334,7 +336,7 @@ def download_rego_img(
     return download_paths
 
 
-def download_rego_skymap(location_code: str, force_download: bool = False) -> List[pathlib.Path]:
+def download_rego_skymap(location_code: str, redownload: bool = False) -> List[pathlib.Path]:
     """
     Download all of the REGO skymap IDL .sav files.
 
@@ -342,7 +344,7 @@ def download_rego_skymap(location_code: str, force_download: bool = False) -> Li
     ----------
     location_code: str
         The ASI station code, i.e. ``ATHA``
-    force_download: bool
+    redownload: bool
         If True, download the file even if it already exists. Useful if a prior
         data download was incomplete.
 
@@ -378,15 +380,15 @@ def download_rego_skymap(location_code: str, force_download: bool = False) -> Li
         skymap_name = utils._search_hrefs(skymap_folder_absolute, search_pattern='.sav')[0]
         skymap_save_name = skymap_name.replace('-%2B', '')  # Replace the unicode '+'.
 
-        # Download if force_download=True or the file does not exist.
+        # Download if redownload=True or the file does not exist.
         download_path = pathlib.Path(save_dir, skymap_save_name)
         download_paths.append(download_path)
-        if force_download or (not download_path.is_file()):
+        if redownload or (not download_path.is_file()):
             utils._stream_large_file(skymap_folder_absolute + skymap_name, download_path)
     return download_paths
 
 
-def _download_one_img_file(asi_array_code, location_code, base_url, time, force_download):
+def _download_one_img_file(asi_array_code, location_code, base_url, time, redownload):
     """
     Download one hour-long file.
     """
@@ -400,177 +402,192 @@ def _download_one_img_file(asi_array_code, location_code, base_url, time, force_
     download_path = pathlib.Path(
         asilib.config['ASI_DATA_DIR'], asi_array_code.lower(), file_names[0]
     )
-    if force_download or (not download_path.is_file()):
+    if redownload or (not download_path.is_file()):
         utils._stream_large_file(server_url, download_path)
     return download_path
 
 #TODO: When cleaning up imager, remove all above
 class Downloader:
-    def __init__(self, base_url:str) -> None:
-        """
-        Handles traversing the directory structure on a server and download files.
-        Parameters
-        ----------
-        base_url: str
-            The base URL for the dataset (excludes directories contaning dates, 
-            for example).
-        Example
-        -------
-        d = Download(
-            'https://data.phys.ucalgary.ca/sort_by_project/THEMIS/asi/stream0/'
-            )
-        d.find_file(
-            ['2014', '05', '05', 'gill*', 'ut05', '20140505_0505_gill*.pgm.gz']
-            )
-        d.download()
-        """
-        self.base_url = base_url
+    """
+    Traverses and lists the directory structure on a server and download files.
 
-        # Check that the server status code is not
-        # between 400-599 (error).
-        r = requests.get(self.base_url)
-        status_code = r.status_code
-        if status_code // 100 in [4, 5]:
-            raise ConnectionError(f'{self.base_url} returned a {status_code} response (error).')
+    Parameters
+    ----------
+    url: str
+        The dataset URL. If it points to a folder it must end with a "/".
+    download_dir: str or pathlib.Path
+        The download directory. Must either specify here, or when you call 
+        Downloader.download().
 
-        if self.base_url[-1] != '/':
-            self.base_url += '/'
+    Example
+    -------
+    | # List all of the SAMPEX-HILT State4 files and download the first one.
+    |
+    | import sampex
+    |
+    | d = sampex.Downloader(
+    |     'https://izw1.caltech.edu/sampex/DataCenter/DATA/HILThires/State4/',
+    |     download_dir=sampex.config['data_dir']
+    | )
+    | paths = d.ls(match='*.txt*')
+    | print(f"The downloaded files are: {paths}")
+    | 
+    | print(f"The first file's name is: {paths[0].name} at url {paths[0].url}")
+    | path = paths[0].download(stream=False)
+    | print(f'The file was downloaded to {path}')
+    """
+    def __init__(self, url:str, download_dir=None) -> None:
+        self.url = url
+        self.download_dir=download_dir
         return
 
-    def find_url(self, subdirectories=[], filename=None):
+    def ls(self, match: str='*') -> List[Downloader]:
         """
-        Descends the directory tree starting from self.base_url using folders
-        specified by path.
+        List files and folders in self.url. 
+
         Parameters
         ----------
-        subdirectories: list
-            A list of strings containing subfolders and a filename at the end.
-            You can use a wildcard character (*), to search a unique subdirectory 
-            or file name. The search must only result in exactly one match or an
-            AssertionError will be raised.
-        filename: str
-            An optional filename pattern (including a possible wildcard) that will
-            be used to search the final subdirectory for urls.
+        match: str
+            An optional string pattern to match.
         
         Return
         ------
         list
             A list of full URLs.
         """
-        self.url = self.base_url
-        assert all([isinstance(i, str) for i in subdirectories]), 'The subdirectories must be all strings.'
+        self._check_url_status()
+        matched_hrefs = self._search_hrefs(self.url, match=match)
+        cls = type(self)
+        downloaders = [None]*len(matched_hrefs)
+        for i, matched_href in enumerate(matched_hrefs):
+            new_url = urllib.parse.urljoin(self.url, matched_href, allow_fragments=True)
+            downloaders[i] = cls(new_url, download_dir=self.download_dir)
+        return downloaders
 
-        for subdirectory in subdirectories:
-            if '*' not in subdirectory:
-                self.url = self.url + subdirectory
-            else:
-                matched_hrefs = self._search_hrefs(self.url, search_pattern=subdirectory)
-                assert len(matched_hrefs) == 1, (f'Only one href is allowed but '
-                    f'{len(matched_hrefs)} were found in {self.url}. matched_hrefs (links)={matched_hrefs}')
-                self.url = self.url + matched_hrefs[0]
-
-            if self.url[-1] != '/':
-                self.url += '/'
-
-        if filename is not None:
-            file_hrefs = self._search_hrefs(self.url, search_pattern=filename)
-            _url = []
-            for file_href in file_hrefs:
-                _url.append(self.url + file_href)
-            self.url = _url
-        else:
-            self.url = [self.url]  # For consistency
-        return self.url
-
-    def download(self, save_dir, overwrite=False) -> None:
+    def download(self, download_dir=None, redownload: bool=False, stream: bool=False) -> pathlib.Path:
         """
-        Downloads (streams) a file from self.url to the save_dir directory. I decided
-        to use streaming so that a large file wont overwhelm the user's RAM.
+        Downloads file from self.url to the download_dir directory.
+
         Parameters
         ----------
-        save_dir: str or pathlib.Path
-            The parent directory where to save the data to.
-        overwrite: bool
-            Will download overwrite an existing file. 
+        download_dir: str or pathlib.Path
+            The parent directory where to save the data to. Set it either
+            here or when initializing the class. This 
+        redownload: bool
+            Will redownload an existing file. 
+        stream: bool
+            Download the data in one chunk if False, or break it up and download it 
+            in 5 Mb chunks if True.
+
         Returns
         -------
-        list
-            A list of local paths where the data was saved to. 
+        pathlib.Path
+            The full path to the file. 
         """
-        save_dir = pathlib.Path(save_dir)
-        save_dir.mkdir(parents=True, exist_ok=True)
-        self.save_path = []
+        self._check_url_status()
+        if download_dir is None and self.download_dir is None:
+            raise ValueError(f'download_dir kwarg needs to be set either '
+                             f'in Downloader() or Downloader.download.')
+        if download_dir is not None:
+            self.download_dir = download_dir
 
-        for url in self.url:
-            save_name = url.split('/')[-1]
-            _save_path = save_dir / save_name
-            self.save_path.append(_save_path)
+        self.download_dir = pathlib.Path(self.download_dir)
+        self.download_dir.mkdir(parents=True, exist_ok=True)
+        file_name = pathlib.Path(self.url).name
+        download_path = self.download_dir / file_name
 
-            if (_save_path.exists()) and (not overwrite):
-                continue
+        if (download_path.exists()) and (not redownload):
+            return download_path
 
-            r = requests.get(url, stream=True)
+        if stream:
+            r = requests.get(self.url, stream=True)
             file_size = int(r.headers.get('content-length'))
             downloaded_bites = 0
 
             megabyte = 1024 * 1024
 
-            with open(_save_path, 'wb') as f:
+            with open(download_path, 'wb') as f:
                 for data in r.iter_content(chunk_size=5*megabyte):
                     f.write(data)
                     # Update the downloaded % in the terminal.
                     downloaded_bites += len(data)
                     download_percent = round(100 * downloaded_bites / file_size)
                     download_str = "#" * (download_percent // 5)
-                    print(f'Downloading {save_name}: |{download_str:<20}| {download_percent}%', end='\r')
+                    print(f'Downloading {file_name}: |{download_str:<20}| {download_percent}%', end='\r')
             print()  # Add a newline
-        return self.save_path
+        else:
+            r = requests.get(self.url)
+            with open(download_path, 'wb') as f:
+                f.write(r.content)
+            print(f'Downloaded {file_name}.')
+        return download_path
 
-
-    def _search_hrefs(self, url: str, search_pattern: str = '.cdf') -> List[str]:
+    @property
+    def name(self):
         """
-        Given a url string, this function returns all of the
-        hyper references (hrefs, or hyperlinks). If search_pattern is not
-        specified, a default '.cdf' value is assumed and this function
-        will return all hrefs with the CDF extension. If no hrefs containing
-        search_pattern are found, this function raises a FileNotFound.
-        The search is case-insensitive.
+        Get the url filename
+        """
+        _url = pathlib.Path(self.url)
+        if _url.suffix != '':
+            # url points to a filename. pathlib is not designed for
+            # urls, but this is the easiest way to get the name.
+            return _url.name
+        else:
+            return None
+
+    def _search_hrefs(self, url: str, match: str='*') -> List[str]:
+        """
+        Given a url string, find all hyper references matching the 
+        string match. The re module does the matching.
+        
         Parameters
         ----------
         url: str
-            A url in string format
-        search_pattern: str (optional)
-            Find the exact search_pattern text contained in the hrefs.
-            By default all hrefs matching the extension ".cdf" are returned.
+            A url
+        match: str (optional)
+            The regex match to compare the hyper references to. The default is
+            to match everything (a wildcard)
+
         Returns
         -------
-        hrefs: List(str)
-            A list of hrefs that contain the search_pattern.
+        List(str)
+            A list of hyper references that matched the match string.
+
         Raises
         ------
         FileNotFoundError
-            If a hyper reference (a folder or a file) is not found on the
-            server. This is raised if the data does not exist.
+            If no hyper references were found.
         """
-        matched_hrefs = []
-
         request = requests.get(url)
-        # request.status_code
         soup = BeautifulSoup(request.content, 'html.parser')
 
-        search_pattern = search_pattern.replace('*', '.*')  # regex wildcard
-        for href in soup.find_all('a', href=True):
-            match = re.search(search_pattern, href['href'])
-            if match:
-                matched_hrefs.append(href['href'])
+        match = match.replace('*', '.*')  # regex wildcard
+        href_objs = soup.find_all('a', href=re.compile(match))
+        # I found "?" to be in the column names so we should ignore them.
+        matched_hrefs = [href_obj['href'] for href_obj in href_objs 
+                    if href_obj['href'].find('?')==-1]
         if len(matched_hrefs) == 0:
             raise FileNotFoundError(
                 f'The url {url} does not contain any hyper '
-                f'references containing the search_pattern="{search_pattern}".'
+                f'references containing the match kwarg="{match}".'
             )
         return matched_hrefs
 
-if __name__ == '__main__':
-    d = Downloader('https://data.phys.ucalgary.ca/sort_by_project/THEMIS/asi/stream0/')
-    d.find_file(['2014', '05', '05', 'gill*', 'ut05', '20140505_0505_gill*.pgm.gz'])
-    d.download() 
+    def _check_url_status(self):
+        """
+        Check that the server status code is not
+        between 400-599 (error).
+        """
+        r = requests.get(self.url)
+        status_code = r.status_code
+        if status_code // 100 in [4, 5]:
+            raise ConnectionError(f'{self.url} returned a {status_code} error response.')
+        return
+
+    def __repr__(self) -> str:
+        params = f'{self.url}, download_dir={self.download_dir},'
+        return f'{self.__class__.__qualname__}(' + params + ')'
+    
+    def __str__(self) -> str:
+        return (f'{self.__class__.__qualname__} with url={self.url} and '
+               f'download_dir={self.download_dir}')