+++ asilib/plot/animate_fisheye.py
@@ -83,7 +83,7 @@ def animate_fisheye_generator(
     asi_array_code: str,
     location_code: str,
     time_range: utils._time_range_type,
-    force_download: bool = False,
+    overwrite: bool = False,
     label: bool = True,
     color_map: str = 'auto',
     color_bounds: Union[List[float], None] = None,
@@ -91,8 +91,7 @@ def animate_fisheye_generator(
     azel_contours: bool = False,
     ax: plt.Axes = None,
     movie_container: str = 'mp4',
-    ffmpeg_output_params={},
-    overwrite: bool = False,
+    ffmpeg_output_params={}
 ) -> Generator[Tuple[datetime, np.ndarray, plt.Axes, matplotlib.image.AxesImage], None, None]:
     """
     A generator function that loads the ASI data and then yields individual ASI images,
@@ -112,9 +111,9 @@ def animate_fisheye_generator(
         The ASI station code, i.e. ``ATHA``
     time_range: list of datetime.datetimes or stings
         Defined the duration of data to download. Must be of length 2.
-    force_download: bool
-        If True, download the file even if it already exists. Useful if a prior
-        data download was incomplete.
+    overwrite: bool
+        If true, the output animation will be overwritten, otherwise it will 
+        prompt the user to answer y/n.
     label: bool
         Flag to add the "asi_array_code/location_code/image_time" text to the plot.
     color_map: str
@@ -139,9 +138,6 @@ def animate_fisheye_generator(
         Sets the 'lin' linear or 'log' logarithmic color normalization.
     azel_contours: bool
         Switch azimuth and elevation contours on or off.
-    overwrite: bool
-        If true, the output will be overwritten automatically. If false it will
-        prompt the user to answer y/n.
 
     Yields
     ------
@@ -182,7 +178,7 @@ def animate_fisheye_generator(
     """
     try:
         image_times, images = load_image(
-            asi_array_code, location_code, time_range=time_range, force_download=force_download
+            asi_array_code, location_code, time_range=time_range
         )
     except AssertionError as err:
         if '0 number of time stamps were found in time_range' in str(err):
@@ -262,7 +258,7 @@ def animate_fisheye_generator(
             )
 
         if azel_contours:
-            _add_azel_contours(asi_array_code, location_code, image_time, ax, force_download)
+            _add_azel_contours(asi_array_code, location_code, image_time, ax)
 
         # Give the user the control of the subplot, image object, and return the image time
         # so that the user can manipulate the image to add, for example, the satellite track.
@@ -301,7 +297,6 @@ def _write_movie(image_paths, movie_save_path, ffmpeg_output_params, overwrite):
         The movie file name.
     overwrite: bool
         Overwrite the movie.
-
     """
     ffmpeg_params = {
         'framerate': 10,
@@ -332,7 +327,6 @@ def _add_azel_contours(
     location_code: str,
     time: utils._time_type,
     ax: plt.Axes,
-    force_download: bool,
     color: str = 'yellow',
 ) -> None:
     """
@@ -348,12 +342,10 @@ def _add_azel_contours(
         Time is used to find the relevant skymap file: file created nearest to, and before, the time.
     ax: plt.Axes
         The subplot that will be drawn on.
-    force_download: bool
-        If True, download the file even if it already exists.
     color: str (optional)
         The contour color.
     """
-    skymap_dict = load_skymap(asi_array_code, location_code, time, force_download=force_download)
+    skymap_dict = load_skymap(asi_array_code, location_code, time)
 
     az_contours = ax.contour(
         skymap_dict['FULL_AZIMUTH'],